pyproject.toml
[project]
name = "voicecommander"
version = "0.2.0"
description = "comenzi vocale in romana pentru diferite dispozitive"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.0.0",
    "sounddevice>=0.5.0",
    "numpy>=1.24.0",
    "google-genai>=1.0.0",
    "djitellopy>=2.5.0",
    "librosa>=0.11.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
